% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{d2l_full}
\alias{d2l_full}
\title{Hessian of the log-likelihood for a combined assessment.}
\usage{
d2l_full(resp, w, parms, theta1, theta2, obs = T, parallel = F)
}
\arguments{
\item{resp}{a data.frame containing the binary item responses of both the individual assessment and the (conjunctively scored) group assessment. See details for information on formatting.}

\item{w}{the weight parameter of the RSC model.}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively. See details for information on formatting.}

\item{theta1}{the latent trait of member 1.}

\item{obs}{logical: should the observed value be returned? If not, the expected value is returned.}

\item{theta1}{the latent trait of member 2.}
}
\value{
\code{3 * K} by \code{3 * K} block-diagonal, symmmetrical matrix of second derivatives, with \code{K = length(w)} and rows/cols ordered as \code{rep(c(w_k, theta1_k, theta2_k), times = K)}.
}
