% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRF_functions.R
\name{d2logL}
\alias{d2logL}
\title{Second derviative of log-likelihood of 2PL, in theta.}
\usage{
d2logL(resp, parms, theta, obs = T)
}
\arguments{
\item{resp}{a matrix or data.frame containing the binary item responses.}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta}{the latent trait.}
}
\value{
\code{length(theta)}-vecor of (possibly weigthed) log-likelihoods.
}
\description{
Used to compute standard errors simultaneously estimating individual and group assessments.
}
