% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{data_gen}
\alias{data_gen}
\title{Generate data from the one-parameter RSC model.}
\usage{
data_gen(n_reps, w, parms, theta1, theta2, theta1_se = NULL,
  theta2_se = NULL, NA_pattern = NULL)
}
\arguments{
\item{n_reps}{integer indicating how many datasets to generate.}

\item{w}{the weight parameter of the RSC model.}

\item{parms}{a list or data.frame with elements parms$alpha and parms$beta corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta1}{the latent trait for member 1.}

\item{theta2}{the latent trait for member 2.}

\item{theta1_se}{the standard error of the latent trait for member 1 (Optional). If included, data generation samples \code{n_reps} values from \code{rnorm(theta1, theta1_se)}.}

\item{theta2_se}{the standard error of the latent trait for member 2 (Optional). If included, data generation samples \code{n_reps} values from \code{rnorm(theta2, theta2_se)}.}

\item{NA_pattern}{an (optional) \code{length(theta1)} by \code{nrow(parms)} data.frame with \code{NA} entries for missing data. The missing values are preserved in the generated data.}
}
\value{
A data.frame with \code{length(theta1)} rows containing an id variable for each pair and each sample, the data generating values of theta1, theta2, and mix_prop; the model used to simulate the response pattern; and the simulated response pattern.
}
\description{
This is a wrapper for code{sim_RSC} that includes the data generating parms.
}
