% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRF_functions.R
\name{dIRF}
\alias{dIRF}
\title{First derivative of 2PL IRF in theta.}
\usage{
dIRF(parms, theta)
}
\arguments{
\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta}{the latent trait.}
}
\value{
\code{length(theta)} by \code{nrow(parms)} matrix of first derivatives of IRF.
}
\description{
Used for obtaining SEs of theta and computing WML.
}
