% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRF_functions.R
\name{dJ}
\alias{dJ}
\title{Derviative of J.}
\usage{
dJ(resp, parms, theta)
}
\arguments{
\item{resp}{a matrix or data.frame containing the binary item responses with \code{NA} for missing responses (used omit missing responses).}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta}{the latent trait.}
}
\value{
\code{length(theta)} vector of derivatives of J.
}
\description{
Computes the derivative of J (for 2PL, second derivative of test info).
}
