% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRF_functions.R
\name{dlogL}
\alias{dlogL}
\title{First derviative of log-likelihood of 2PL, in theta.}
\usage{
dlogL(resp, parms, theta)
}
\arguments{
\item{resp}{a matrix or data.frame containing the binary item responses.}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta}{the latent trait.}
}
\value{
\code{length(theta)}- vector of first derivatives of log-likelihood of 2PL, in theta.
}
\description{
Used to estimate theta in \code{est_2PL}.
}
