% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{dlp}
\alias{dlp}
\title{Gradient of the log of the prior distribution for a combined assessment.}
\usage{
dlp(u, theta1, theta2, sigma = 3)
}
\arguments{
\item{u}{the logit of the weight parameter of the RSC model.}

\item{theta1}{the latent trait of member 1.}

\item{theta2}{the latent trait of member 2.}

\item{sigma}{prior standard deviation for logit of weight.}
}
\value{
\code{3 * K} - vector of first derivatives, with \code{K = length(u)}, ordered as \code{rep(c(u_k, theta1_k, theta2_k), times = K)}.
}
\description{
This function computes the derivative of the log of the prior for a combined assessment, in which the 2PL model is used for the individual component of the assessment and the one-parameter RSC model is used for the group component of the assessment. A standard normal prior is used for individual ability, and for the logit of the weight of the RSC model a normal prior is used with standard deviation \code{sigma}.
}
