% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRF_functions.R
\name{est_2PL}
\alias{est_2PL}
\title{Estimation of latent trait for 2PL model.}
\usage{
est_2PL(resp, parms, method = "ML", parallel = T)
}
\arguments{
\item{resp}{a matrix or data.frame containing the binary item responses.}

\item{parms}{a list or data.frame with elements parms$alpha and parms$beta corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{method}{one of \code{"ML", "WML", "MAP"}.}

\item{logical:}{call \code{parallel:mclapply} instead of looping over \code{nrow(resp)}?}
}
\value{
An \code{nrow(resp)} by 3 matrix of loglikleihoods, theta estimates, analytic SEs for each response pattern in resp.
}
\description{
The latent trait (theta) is estimated by calling \code{uniroot} on q = dlogL + C, where dlogL is the first derivative of the log-likelihood, in theta, and C is a function of theta that depends on the method selected. For \code{method = "ML"}, C = 0. For \code{method = "WML"}, C is given by equation 9 of Warm (1989). For \code{method = "MAP"}, C = - theta. Standard errors (or posterior standard deviations) are computed analytically via the test information of each estimator. The value of logL (not the value of q) at the estimate is also provided. If \code{parallel = T}, the call to \code{uniroot} is parallelized via \code{parallel::mclapply}, but the decrease in runtime is negligible for \code{dim(resp) < c(1000, 100)}
}
\details{
ToDo: 1. Remove and index duplicate response patterns. 2. Find a better way of setting the range of x in \code{uniroot}.
}
