% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{est_RSC2}
\alias{est_RSC2}
\title{Estimation of the one-parameter RSC model, with latent traits assumed to be known.}
\usage{
est_RSC2(resp, parms, theta1, theta2, method = "MAP", obs = F,
  epsilon = 0.05, parallel = F)
}
\arguments{
\item{resp}{a matrix or data.frame containing the (conjunctively-scored) binary item responses.}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta1}{the latent trait of member 1.}

\item{method}{one of \code{c("ML", "MAP")}. The latter is strongly recommended.}

\item{epsilon}{a small positive number, see description for details.}

\item{parallel}{logical: call \code{parallel:mclapply} instead of looping over \code{nrow(resp)}?}

\item{theta1}{the latent trait of member 2.}
}
\value{
An named \code{nrow(resp)} by 3 data.frame containing the estimates, their standard errors, and the value of the log-likelihood of the RSC model at the solution (not log posterior with MAP).
}
