% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{format_NA}
\alias{format_NA}
\title{Transforms one matrix/data.frame have the same NA entries as another matrix/data.frame}
\usage{
format_NA(data, NA_pattern = NULL)
}
\arguments{
\item{data}{the matrix/data.frame into which to write NA entries.}

\item{NA_pattern}{the matrix/data.frame from which to write NA entries.}
}
\value{
a data.frame/matrix similar to \code{data} that includes all the NA entires as \code{NA_pattern}. If \code{NA_pattern == NULL}, the data.frame/matrix is returned unaltered.
}
\description{
Useful for reproducing observed NA patterns in simulated data. If \code{q = (nrow(data) / nrow(NA_pattern)} is an integer greater than 1, will replicate NA_pattern as \code{kronecker(as.matrix(NA_pattern), rep(1, q)}.
}
