% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{format_resp}
\alias{format_resp}
\title{Formats responses a response df to match the calibration df}
\usage{
format_resp(resp, calib, version = NULL)
}
\arguments{
\item{resp}{the target df to be formatted}

\item{calib}{the calibration df}

\item{version}{an optional string used to subset resp via \code{grep(version, names(resp))}}
}
\value{
a version of resp that has the same names as calib
}
\description{
Drops items in the target df not in the calibration sample; adds items (with NA entries) in the calibration sample not in the target df. Used for obtaining factor scores from calibration sample \code{ltm}.
}

