% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Ind}
\alias{Ind}
\title{Item response function for ``the Independence model"}
\usage{
Ind(alpha, beta, theta1, theta2)
}
\arguments{
\item{alpha}{the item discriminations}

\item{beta}{the item difficulties}

\item{theta1}{the latent trait for member 1}

\item{theta2}{the latent trait for member 2}
}
\value{
\code{length(theta)} by \code{length(alpha)} matrix of response probabilities
}
\description{
Computes a matrix of probabilities for correct responses using the independence model of collaboration for the members and the 2PL model for the items.
}

