% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{Info_u}
\alias{Info_u}
\title{Item info: u.}
\usage{
Info_u(u, parms, theta1, theta2)
}
\arguments{
\item{u}{is the logit of the weight parameter of the RSC model.}

\item{parms}{a data.frame with columns \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta1}{the latent trait of member 1.}

\item{theta2}{the latent trait of member 2.}
}
\value{
\code{length(w)} by \code{nrow(parms)} matrix of item info.
}
\description{
Computes item info for the logit of the one-parameter RSC weight.
}
