% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRF_functions.R
\name{IRF}
\alias{IRF}
\title{The IRF for the two parameter logistic model.}
\usage{
IRF(parms, theta)
}
\arguments{
\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta}{the latent trait.}
}
\value{
\code{length(theta)} by \code{nrow(parms)} matrix of response probabilities.
}
\description{
Computes a matrix of probabilities for correct responses under 2PL model.
}
