% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{l_full_sum}
\alias{l_full_sum}
\title{Log-likelihood of a combined assessment, with sum option.}
\usage{
l_full_sum(resp, w, parms, theta1, theta2, Sum = F)
}
\arguments{
\item{resp}{a data.frame containing the binary item responses of both the individual assessment and the (conjunctively scored) group assessment. See details for information on formatting.}

\item{w}{the weight parameter of the RSC model.}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively. See details for information on formatting.}

\item{theta1}{the latent trait of member 1.}

\item{theta2}{the latent trait of member 2.}
}
\value{
\code{length(w)}-vector of log-likelihoods.
}
\description{
This function is identical to \code{l_full}, except that it has an option for whether or not to sum the log-liklihood over respondents. See \code{help(l_full)} for details on formatting \code{resp} and \code{parms}.
}
