% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logL}
\alias{logL}
\title{Log-ikelihood of given matrix of binary responses for a stated model, conditional on theta.}
\usage{
logL(resp, model, alpha, beta, theta1, theta2 = NULL)
}
\arguments{
\item{resp}{the matrix binary responses}

\item{model}{is one of \code{c("twoPL", "Ind", "Min", "Max", "AI") }}

\item{alpha}{the item discriminations}

\item{beta}{the item difficulties}

\item{theta1}{the latent trait for member 1}

\item{theta2}{the latent trait for member 2}
}
\value{
An \code{nrow(resp)}-vector of log-likleihoods for each response pattern.
}
\description{
Computes a vector of likelihoods for a given matrix binary response patterns, for a selection of models for pairwise collaboration, conditional on theta
}

