% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{lp}
\alias{lp}
\title{Log of prior distribution for a combined assessment.}
\usage{
lp(u, theta1, theta2, sigma = 3)
}
\arguments{
\item{u}{the logit of the weight parameter of the RSC model.}

\item{theta1}{the latent trait of member 1.}

\item{theta2}{the latent trait of member 2.}

\item{sigma}{prior standard deviation for logit of weight.}
}
\value{
\code{length(u)}-vector of log-priors (minus a constant).
}
\description{
This function computes the log of the prior (minus a constant) for a combined assessment, in which the 2PL model is used for the individual component of the assessment and the one-parameter RSC model is used for the group component of the assessment. A standard normal prior is used for individual ability, and for the logit of the weight of the RSC model a normal prior is used with standard deviation \code{sigma}.
}
