% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lr_test}
\alias{lr_test}
\title{Likelihood ratio tests for various models of collaboration.}
\usage{
lr_test(resp, model, alpha, beta, ind_theta, col_theta, n_boot = 0)
}
\arguments{
\item{resp}{the matrix binary data from the \strong{collaborative responses}}

\item{model}{is one or more of \code{c("Ind", "Min", "Max", "AI") }}

\item{alpha}{the item discriminations of the resp items}

\item{beta}{the item difficulties of the resp items}

\item{ind_theta}{the latent trait for each member, as estimated from a non-collaborative form}

\item{col_theta}{the latent trait for each member, as estimated from a collaborative form}

\item{n_boot}{number of bootstraps to use for testing the likelihood ratio.}
}
\value{
A \code{length(ind_theta)/2} by \code{length(model)} matrix, each column of which contains the lr_tests for each model.
}
\description{
Computes a likelihood ratio test for one or more models of pairwise collaboration, given ``assumed to be known" item and person parameters (i.e., neither estimation error in item parameters nor prediction error in latent variables is accounted for by this procedure).
}

