% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRF_functions.R
\name{M}
\alias{M}
\title{Computes mutiplier for first deriviative of log-likelihood of 2PL.}
\usage{
M(resp, parms, theta)
}
\arguments{
\item{resp}{a matrix or data.frame containing the binary item responses.}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta}{the latent trait.}
}
\value{
\code{dim(resp)}-matrix of multipliers.
}
\description{
Writing the derivative of the log-likelihood of the 2PL in theta for a single item as M * dIRF, this function computes (a matrix of values of) M, with dIRF being the first deriviative of the 2PL IRF. Called by functions that compute derivative of the log-likelihood.
}
