% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ml_twoPL}
\alias{ml_twoPL}
\title{Internal function used in lr_test; under development.}
\usage{
ml_twoPL(resp, alpha, beta, method = "ML")
}
\description{
Internal function used in lr_test; under development.
}

