% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{Mstar}
\alias{Mstar}
\title{Mutiplier term for first deriviative of log-likelihood of one-parameter RSC model.}
\usage{
Mstar(resp, u, parms, theta1, theta2)
}
\arguments{
\item{resp}{a matrix or data.frame containing the (conjunctively-scored,  binary) item responses.}

\item{u}{the logit of the weight parameter of the RSC model.}

\item{parms}{a data.frame with columns \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta1}{the latent trait of member 1.}

\item{theta2}{the latent trait of member 2.}
}
\value{
\code{dim(resp)}-matrix of multipliers.
}
\description{
The gradient of the log-likelihood of the RSC model a single item can be computed as M * grad_IRF. This function computes (a matrix of values of) M. Called by functions that compute derivatives of the log-likelihood.
}
