% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRF_functions.R
\name{N}
\alias{N}
\title{Multiplier for second deriviative of log-likelihood of 2PL.}
\usage{
N(resp, parms, theta, obs = F)
}
\arguments{
\item{resp}{a matrix or data.frame containing the binary item responses.}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta}{the latent trait.}

\item{obs}{logical: should the observed value be returned? If not, the expected value is returned.}
}
\value{
\code{dim(resp)}-matrix of multipliers.
}
\description{
Similar to \code{M}, but for the second derivatives. Called by functions that compute second derivative of the log-likelihood.
}
