% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{neg_logL}
\alias{neg_logL}
\title{Internal function used in lr_test; under development.}
\usage{
neg_logL(theta, resp, alpha, beta)
}
\description{
Internal function used in lr_test; under development.
}

