% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRF_functions.R
\name{sim_2PL}
\alias{sim_2PL}
\title{Simulate item responses from the 2PL.}
\usage{
sim_2PL(parms, theta)
}
\arguments{
\item{parms}{a list or data.frame with elements parms$alpha and parms$beta corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta}{the vector of latent traits.}
}
\value{
\code{length(theta)} by \code{nrow(parms)} matrix of binary responses.
}
\description{
Simulate item responses from the 2PL.
}
