% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data from a model of pairwise collaboration.}
\usage{
sim_data(model, alpha, beta, theta1 = 0, theta2 = 0)
}
\arguments{
\item{model}{is one of \code{c("twoPL", "Ind", "Min", "Max", "AI") }}

\item{alpha}{the item discriminations}

\item{beta}{the item difficulties}

\item{theta1}{the latent trait for member 1}

\item{theta2}{the latent trait for member 2}
}
\value{
\code{length(theta1)} by \code{length(alpha)} matrix of binary response patterns
}
\description{
Simulate data using either the 2PL or a model for pariwise collaboration obtained from the 2PL.
}

