% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{sim_RSC}
\alias{sim_RSC}
\title{Simulate item responses from the one-parameter RSC model.}
\usage{
sim_RSC(w, parms, theta1 = 0, theta2 = 0)
}
\arguments{
\item{w}{the weight parameter of the RSC model.}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively. See details for information on formatting.}

\item{theta1}{the latent trait of member 1.}

\item{theta1}{the latent trait of member 2.}
}
\value{
\code{length(theta1)} by \code{nrow(parms)} matrix of binary responses.
}
\description{
Simulate item responses from the one-parameter RSC model.
}
