% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_twoPL}
\alias{sim_twoPL}
\title{Simulate data from 2PL model.}
\usage{
sim_twoPL(alpha, beta, theta = 0)
}
\arguments{
\item{alpha}{the item discriminations}

\item{beta}{the item difficulties}

\item{theta}{the values of latent trait to simulate data for}
}
\value{
\code{length(theta)} by \code{length(alpha)} matrix of binary response patterns
}
\description{
Simulate data using the 2PL model.
}

