% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{theta_gen}
\alias{theta_gen}
\title{Generates or replicates values of latent trait}
\usage{
theta_gen(n, theta, theta_se = NULL)
}
\arguments{
\item{n}{number of values to generate / replicate for each value of \code{theta}.}

\item{theta}{the latent trait.}

\item{theta_se}{the standard error of the latent trait.}
}
\value{
An \code{n * length(theta)} vector in which each value \code{theta} is generated / replicated \code{n} times.
}
\description{
If \code{theta_se}' is not null, values are generated using \code{rnorm}. Otherwise, each value is replicated.
}
