% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{twoPL}
\alias{twoPL}
\title{Item response function for 2PL}
\usage{
twoPL(alpha, beta, theta)
}
\arguments{
\item{alpha}{the item discriminations}

\item{beta}{the item difficulties}

\item{theta}{the latent trait}
}
\value{
\code{length(theta)} by \code{length(alpha)} matrix of response probabilities
}
\description{
Computes a matrix of probabilities for correct responses under 2PL model
}

