% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{var_rsc}
\alias{var_rsc}
\title{Computes variance of latent traits and the one-parameter RSC model for a combined assessment.
#' @param resp a data.frame containing the binary item responses of both the individual assessment and the (conjunctively scored) group assessment. See details of \code{rsc} for information on formatting.}
\usage{
var_rsc(resp, u, parms, theta1, theta2, method = "MAP", obs = F,
  sigma = 3)
}
\arguments{
\item{u}{the logit of the weight parameter of the RSC model.}

\item{parms}{a data.frame with columns \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively. See details of \code{rsc} for information on formatting.}

\item{theta1}{the latent trait of member 1.}

\item{theta2}{the latent trait of member 2.}

\item{method}{one of \code{c("ML", "MAP")}. The latter isrecommended.}

\item{obs}{logical: should standard errors be computed using the observed (\code{TRUE}) or expected (\code{FALSE}) Hessian?}

\item{sigma}{prior standard deviation for logit of weight.}
}
\value{
\code{3 * K} by \code{3 * K} covariance matrix of the parameter estimates, with \code{K = length(u)} and rows/cols ordered as \code{rep(c(u_k, theta1_k, theta2_k), times = K)}.
}
\description{
Computes variance of latent traits and the one-parameter RSC model for a combined assessment.
#' @param resp a data.frame containing the binary item responses of both the individual assessment and the (conjunctively scored) group assessment. See details of \code{rsc} for information on formatting.
}
