% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{dl_full}
\alias{dl_full}
\title{Gradient of the log-likelihood of a combined assessment.}
\usage{
dl_full(resp, w, parms, theta1, theta2)
}
\arguments{
\item{resp}{a data.frame containing the binary item responses of both the individual assessment and the (conjunctively scored) group assessment. See details for information on formatting.}

\item{w}{the weight parameter of the RSC model.}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively. See details for information on formatting.}

\item{theta1}{the latent trait of member 1.}

\item{theta2}{the latent trait of member 2.}
}
\value{
\code{3 * K} - vector of first derivatives, with \code{K = length(w)}, ordered as \code{rep(c(w_k, theta1_k, theta2_k), times = K)}.
}
