% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{dl_RSC}
\alias{dl_RSC}
\title{Gradient of log-likelihood of one-parameter RSC model.}
\usage{
dl_RSC(resp, w, parms, theta1, theta2)
}
\arguments{
\item{resp}{a matrix or data.frame containing the (conjunctively-scored) binary item responses.}

\item{w}{the weight parameter of the RSC model.}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta1}{the latent trait of member 1.}

\item{theta1}{the latent trait of member 2.}
}
\value{
\code{3 * K} - vector of first derivatives, with \code{K = length(w)}, ordered as \code{rep(c(w_k, theta1_k, theta2_k), times = K)}.
}
