% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{dlp}
\alias{dlp}
\title{Graident of the log of the prior distribution for a combined assessment.}
\usage{
dlp(w, theta1, theta2, epsilon = 0.05)
}
\arguments{
\item{w}{the weight parameter of the RSC model.}

\item{theta1}{the latent trait of member 1.}

\item{theta2}{the latent trait of member 2.}

\item{epsilon}{a small positive number, see description for details.}
}
\value{
\code{3 * K} - vector of first derivatives, with \code{K = length(w)}, ordered as \code{rep(c(w_k, theta1_k, theta2_k), times = K)}.
}
\description{
This function computes the derivative of the log of the prior for a combined assessment, in which the 2PL model is used for the individual component of the assessment and the one-parameter RSC model is used for the group component of the assessment. A standard normal prior is used for individual ability. A two-parameter Beta prior is the parameter of the RSC model, in which both parameters are equal to 1 + epsilon.
}
