% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{est_RSC}
\alias{est_RSC}
\title{Simultaneous estimation of latent traits and the one-parameter RSC model for a combined assessment.}
\usage{
est_RSC(resp, parms, starts = NULL, method = "MAP", obs = F,
  epsilon = 0.05, parallel = F)
}
\arguments{
\item{resp}{a data.frame containing the binary item responses of both the individual assessment and the (conjunctively scored) group assessment. See details for information on formatting.}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively. See details for information on formatting.}

\item{starts}{starting values, ordered as triplets of \code{c(w, theta1, theta2)} for each row or \code{resp} (optional).}

\item{method}{one of \code{c("ML", "MAP")}. The latter is strongly recommended.}

\item{obs}{logical: should standard errors be computed using the observed (\code{TRUE}) or expected (\code{FALSE}) Hessian?}

\item{epsilon}{a small positive number, see description for details.}

\item{parallel}{logical: call \code{parallel:mclapply} instead of looping over \code{nrow(resp)}?}
}
\value{
An named \code{nrow(resp)} by 7 data.frame containing the estimates, their standard errors, and the value of the objective function at the solution.
}
\description{
This function calls \code{optim} to estimate the parameter vector \code{c(w, theta1, theta2)} from the repsonses to a combined assessment, in which the 2PL model is used for the individual component of the assessment and the one-parameter RSC model is used for the group component of the assessment.
}
\details{
Esimation is via either maximum likelihood (ML) or modal a'posteriori (MAP), with the latter being prefered. For MAP, a standard normal prior is used for individual ability. A two-parameter Beta prior is the parameter of the RSC model, in which both parameters are equal to 1 + \code{epsilon}. Standard errors (or posterior standard deviations) are computed by numerically inverting the analytically computed Hessian of the objective function, at the parameter estimates. The value of the objective function at the estimate is is also provided. If \code{parallel = T}, the call to \code{optim} is parallelized via \code{parallel::mclapply}.

The response matrix \code{resp} must be formatted to contain one row of binary responses for each respondent (not each dyad). Members of the same dyad must be on adjancent rows, such that \code{resp[odd,]} gives the responses of one member of a dyad and \code{resp[odd + 1, ]} gives the responses of the other member of the dyad, where \code{odd} is any odd integer in \code{c(1, nrow(resp))}. The (column) names for items on the individual assessment must include \code{"IND"}; those on the (conjunctively-scored) group assessment just include \code{"COL"} -- these text-keys are grepped from \code{names(resp)} to obtain the response patterns for the individual assessment and the group assessment. Note that only the odd rows of \code{resp[grep("COL", names(resp))]} are used when computing the log-likelihood for the group component.

The order of items (columns) of \code{resp} is assumed to correpond to that of items (rows) of \code{parms}, for each of \code{c("IND", "COL")}. Similarly to the procedure described for \code{names(resp)}, \code{row.names(parms)} is grepped for each of \code{c("IND", "COL")} to obtain the item parameters of the individual assessment and the group assessment.
Type \code{l_full} for an illustration of how the formatting calls are made.
}
