% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{format_resp}
\alias{format_resp}
\title{Formats a data.frame to have the same variables (column names) as that of another data.frame}
\usage{
format_resp(resp, items, version = NULL)
}
\arguments{
\item{resp}{the target data.frame to be re-formatted.}

\item{version}{an optional string used to subset resp via \code{grep(version, names(resp))}}

\item{item}{string vector of variable names.}
}
\value{
a data.frame that results from padding and deleting \code{resp} as described.
}
\description{
Drops columns in the target data.frame (resp) not in the variable name list (items). Adds columns (with NA entries) for entries in the items that are not alrady in resp. This is mainly to simplify using item parameters obtained from one sample (e.g., calibration sample) with a second sample that may not use all of the items.
}
