% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{Nstar}
\alias{Nstar}
\title{Multiplier for second deriviative of log-likelihood of one-parameter RSC model.}
\usage{
Nstar(resp, w, parms, theta1, theta2, obs = F)
}
\arguments{
\item{resp}{a matrix or data.frame containing the (conjunctively-scored) binary item responses.}

\item{w}{the weight parameter of the RSC model.}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta1}{the latent trait of member 1.}

\item{theta2}{the latent trait of member 2.}

\item{obs}{logical: should the observed value be returned? If not, the expected value is returned.}
}
\value{
\code{dim(resp)}-matrix of multipliers.
}
\description{
Similar to \code{Mstar}, but for the second derivatives. Called by functions that compute second derivative of the log-likelihood.
}
