% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRF_functions.R
\name{RSC}
\alias{RSC}
\title{The IRF of the one-parameter RSC model.}
\usage{
RSC(w, parms, theta1, theta2)
}
\arguments{
\item{w}{the weight parameter of the RSC model.}

\item{parms}{a named list or data.frame with elements \code{parms$alpha} and \code{parms$beta} corresponding to the discrimination and difficulty parameters of the 2PL model, respectively.}

\item{theta1}{the latent trait of member 1.}

\item{theta2}{the latent trait of member 2.}
}
\value{
\code{length(w)} by \code{nrow(parms)} matrix of response probabilities.
}
\description{
Computes a matrix of probabilities for correct responses under one parameter RSC model, using 2PL model for individual responses.
}
